Organize code in packages.
