+++ .gitignore
@@ -1,2 +1,3 @@
 main
 statsd-librato
+statsd-librato-go
